kb_airflow/requirements.txt
apache-airflow==2.7.2
pyjokes==0.6.0
psycopg2-binary==2.9.9
setuptools==68.2.2
tabulate==0.9.0
Faker==19.10.0
pandas==2.0.0
python-dotenv==1.0.0
